+++ src/app/code/JetRails/Varnish/Console/Command/Status/Set.php
@@ -12,7 +12,7 @@
 	 * Set.php - This class inherits from the AbstractCommand.  This command takes in an additional
 	 * argument, either enable or disable, and sets the caching application to 'Varnish Cache' or
 	 * 'Built-in Cache' respectfully.
-	 * @version         1.1.6
+	 * @version         1.1.7
 	 * @package         JetRails® Varnish
 	 * @category        Status
 	 * @author          Rafael Grigorian - JetRails®
@@ -31,7 +31,7 @@
 			// Register the command and set the arguments
 			$this->setName ("varnish:status:set")
 			->setDescription ("Set what caching application is being used")
-			->addArgument ( "state", InputArgument::REQUIRED, "Should we enable varnish cache?" );
+			->addArgument ( "state", InputArgument::REQUIRED, "Should we enable Varnish Cache™?" );
 		}
 
 		/**