+++ src/app/code/JetRails/Varnish/Controller/Adminhtml/Purge/Url.php
@@ -12,7 +12,7 @@
 	/**
 	 * Url.php - This class is a controller action and when it is triggered, it is responsible for
 	 * purging all the cache in all the configured cache servers that exactly match the passed url.
-	 * @version         1.1.6
+	 * @version         1.1.7
 	 * @package         JetRails® Varnish
 	 * @category        Purge
 	 * @author          Rafael Grigorian - JetRails®
@@ -104,7 +104,7 @@
 			else {
 				// Cache application is not Varnish, warn user
 				$this->messageManager->addError (
-					"Cache application must be set to <b>Varnish Cache</b>, set it by configuring" .
+					"Cache application must be set to <b>Varnish Cache™</b>, set it by configuring" .
 					" <b>Stores → Advanced → Developer → System → Full Page Cache → Caching Application</b>"
 				);
 			}