+++ src/app/code/JetRails/Varnish/Observer/AutoPurge/Page.php
@@ -8,7 +8,7 @@
 	/**
 	 * Page.php - This observer is triggered when the CMS page save event is fired.  It then finds
 	 * the url of the CMS page and sends a URL purge request to the configured varnish servers.
-	 * @version         1.1.6
+	 * @version         1.1.7
 	 * @package         JetRails® Varnish
 	 * @category        Save
 	 * @author          Rafael Grigorian - JetRails®